Initial Commit
Forked and cloned repository from Techstars github account, intalled
yarn using homebrew on Mac, installed react-bootstrap and react-router-
dom. Bootstrap will be used for the creation of the form and so far for
the Navbar. Created Interactive UI using CSS. Will probably continue
styling with styled components instead of the CSS file.
Next step will be the creation of an event and the sign up form.