Changed a approached on this challenge
validateForm.js contains the setup for form validation, Signup form was
originally built using React Hooks and the Form itself was built using
react-bootstrap. Since this is a dummy website with a dummy form I was
having trouble with useEffect() from hooks.I decided not to use
React.Component, for two reasons, number one because it is my personal
preference to use React Hooks due to its simplicity and because it
decreases the size of the code making it easily maintainable and because
I decided to dedicate more time on the design of the page since it is
for a front-end role. Because of that a package was used to create the
form, called react-hooks-form. I believe in using the available tools,
packages, libraries to facilite coding. Long story short, I think a
Signup Component or Hook is unnecessary to complete this challenge. To
follow this challenge's requirements, I have included a file called
FormExample.js which is an example of the same form using the
validateForm.js and submitForm.js files with React.Component but it
will not be included on App.js.