This is a final production commit
will check for spelling errors and bugs and will commit again if needed.
If not found, this is the final commit. Material UI package was
installed but not used. FormExample.js is a dummy React Component that
is missing the form element but works in case needed, as stated on the
previous commit, react-hook-form was used for the form implementation
instead of a regular React Component/Hook. Adjusted the design to all
screen sizes, mobile friendly. Footer link takes you to Techstars page
and my personal portfolio. The links on the page, Start it up, Location,
Sign Up, Why you and Why me are (obviously) a dummy example of how
things would show on the page in case there was a back-end and I could
perform axios calls. They work using a hide/show hook, clicking on
the same link twice will hide or show the same link. That is not a bug.
I wanted to make the page look as close as the example page honoring
Techstars design choices. The text used on the page is not fake, feel
free to read it. App.scss should compile to CSS with no problems.
Styled Components was heavily used on Event.js and somewhat used on
other files. Tests with jest were not performed, honestly because I
never had to (always worked with testing developers). I did research the
testing and will be willing to learn it. Sign.js is the same as
FormExample.js but in Hooks however, that Component is not finished
(stopped working on it to implement react-hook-form).